% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollections.R
\name{getCollections}
\alias{getCollections}
\title{List of available collections}
\usage{
getCollections(db)
}
\arguments{
\item{db}{portal name, for an overview see \code{\link{portals}}}
}
\description{
List of available collections
}
\details{
Get list of available collections from the MyCoPortal. For details also see \url{http://mycoportal.org/portal/collections/index.php}
}
\examples{
\dontrun{
getCollections(db = "SCAN")
}
}
\author{
Franz-Sebastian Krah
}
