% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_onclick_next_page.R
\name{is_onclick_next_page}
\alias{is_onclick_next_page}
\title{Test if consecutive table page reached by onclick function or direct link}
\usage{
is_onclick_next_page(remdriver)
}
\description{
Test if consecutive table page reached by onclick function or direct link
}
