% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portals.R
\name{portal}
\alias{portal}
\title{Find URL for portal db}
\usage{
portal(db = "lichen")
}
\arguments{
\item{db}{character string specifying Symbiota database}
}
\description{
Find URL for portal db
}
\examples{
\dontrun{
portal(db = "lichen")
}
}
\author{
Franz-Sebastian Krah
}
