% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portals.R
\name{portals}
\alias{portals}
\title{Portals available through Symbiota}
\usage{
portals()
}
\description{
Portals available through Symbiota
}
\details{
Symbiota is an open source content management system for curating specimen- and observation-based biodiversity data
}
\examples{
\dontrun{
portals()
}
}
\author{
Franz-Sebastian Krah
}
