% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{records}
\alias{records}
\title{Create a records objects}
\usage{
records(nr.records, citation, query, records, db)
}
\arguments{
\item{nr.records}{A numeric giving the number of records retrieved}

\item{citation}{A character string with the recommended citation from the website}

\item{query}{A list of the user arguments used}

\item{records}{A data.frame with the query records results}

\item{db}{A character string specifying the database}
}
\description{
Create a records objects
}
\examples{
\dontrun{
help("records-class")
}
}
\author{
Franz-Sebastian Krah
}
