% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordsTable.R
\docType{methods}
\name{recordsTable,records-method}
\alias{recordsTable,records-method}
\alias{recordsTable}
\title{Extract records table from class \code{records}}
\usage{
\S4method{recordsTable}{records}(object)
}
\arguments{
\item{object}{object of class \code{records}}
}
\value{
data.frame
}
\description{
Extract records table from class \code{records}
}
\author{
Franz-Sebastian Krah
}
