% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stop_docker.R
\name{start_docker}
\alias{start_docker}
\title{Start Docker}
\usage{
start_docker(verbose = TRUE, wait = 2)
}
\arguments{
\item{verbose}{logical}

\item{wait}{Waiting time for system call to finish}
}
\description{
Start Docker
}
\details{
Docker available for download at: https://www.docker.com
}
\examples{
\dontrun{
## Usually these functions are only used internally.
## However, if there are errors in the function leading to problems running 'symbiota' again
## it might be neccessary to stop_docker.
start_docker() # starts docker
stop_docker() # stops docker
}
}
